core/pyproject.toml
[tool.poetry]
name = "iei-code"
version = "0.1.0"
description = ""
authors = ["SrIzan <izan.garciag@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
attr = "0.3.2"
colorama = "0.4.6"
configparser = "7.1.0"
cryptography = "44.0.0"
docutils = "0.21.2"
fastapi = "0.115.6"
filelock = "3.16.1"
htmlparser = "0.0.2"
ipython = "8.12.3"
ipywidgets = "8.1.5"
keyring = "24.3.1"
numpy = "2.2.0"
pandas = "2.2.3"
pillow = "11.0.0"
protobuf = "5.29.1"
pyopenssl = "24.3.0"
python-dotenv = "1.0.1"
redis = "5.2.1"
selenium = "4.27.1"
sphinx = "8.1.3"
thread = "2.0.5"
urllib3-secure-extra = "0.1.0"
uvicorn = "0.21.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
